Move serialization to private path
