+++ src/model_signing/_cli.py
@@ -24,7 +24,7 @@ import model_signing
 from model_signing import model
 from model_signing._hashing import file
 from model_signing._hashing import memory
-from model_signing.serialization import serialize_by_file
+from model_signing._serialization import serialize_by_file
 from model_signing.signature import key
 from model_signing.signature import pki
 from model_signing.signing import in_toto
