+++ src/model_signing/_serialization/serialization.py
