+++ src/model_signing/_serialization/serialize_by_file_shard.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Model serializers that operated at file level granularity."""
+"""Model serializers that operated at file shard level of granularity."""
 
 import abc
 import base64
@@ -26,48 +26,17 @@ from typing_extensions import override
 
 from model_signing import manifest
 from model_signing._hashing import file
-from model_signing.serialization import serialization
+from model_signing._serialization import serialization
+from model_signing._serialization import serialize_by_file
 
 
-def check_file_or_directory(
-    path: pathlib.Path, *, allow_symlinks: bool = False
-) -> None:
-    """Checks that the given path is either a file or a directory.
-
-    There is no support for sockets, pipes, or any other operating system
-    concept abstracted as a file.
-
-    Furthermore, this would raise if the path is a broken symlink, if it doesn't
-    exists or if there are permission errors.
-
-    Args:
-        path: The path to check.
-        allow_symlinks: Controls whether symbolic links are included. If a
-          symlink is present but the flag is `False` (default) the
-          serialization would raise an error.
-
-    Raises:
-        ValueError: The path is neither a file or a directory, or the path
-          is a symlink and `allow_symlinks` is false.
-    """
-    if not allow_symlinks and path.is_symlink():
-        raise ValueError(
-            f"Cannot use '{path}' because it is a symlink. This"
-            " behavior can be changed with `allow_symlinks`."
-        )
-    if not (path.is_file() or path.is_dir()):
-        raise ValueError(
-            f"Cannot use '{path}' as file or directory. It could be a"
-            " special file, it could be missing, or there might be a"
-            " permission issue."
-        )
-
-
-def _build_header(*, entry_name: str) -> bytes:
-    """Builds a header to encode a path with given name.
+def _build_header(*, name: str, start: int, end: int) -> bytes:
+    """Builds a header to encode a path with given name and shard range.
 
     Args:
         entry_name: The name of the entry to build the header for.
+        start: Offset for the start of the path shard.
+        end: Offset for the end of the path shard.
 
     Returns:
         A sequence of bytes that encodes all arguments as a sequence of UTF-8
@@ -75,37 +44,50 @@ def _build_header(*, entry_name: str) -> bytes:
         dot (so the file digest can be appended unambiguously).
     """
     # Prevent confusion if name has a "." inside by encoding to base64.
-    encoded_name = base64.b64encode(entry_name.encode("utf-8"))
+    encoded_name = base64.b64encode(name.encode("utf-8"))
+    encoded_range = f"{start}-{end}".encode("utf-8")
     # Note: empty string at the end, to terminate header with a "."
-    return b".".join([encoded_name, b""])
+    return b".".join([encoded_name, encoded_range, b""])
 
 
-def _ignored(path: pathlib.Path, ignore_paths: Iterable[pathlib.Path]) -> bool:
-    """Determines if the provided path should be ignored.
+def _endpoints(step: int, end: int) -> Iterable[int]:
+    """Yields numbers from `step` to `end` inclusive, spaced by `step`.
 
-    Args:
-        path: The path to check.
-        ignore_paths: The paths to ignore while serializing a model.
+    Last value is always equal to `end`, even when `end` is not a multiple of
+    `step`. There is always a value returned.
 
-    Returns:
-        Whether or not the provided path should be ignored.
+    Examples:
+    ```python
+    >>> list(_endpoints(2, 8))
+    [2, 4, 6, 8]
+    >>> list(_endpoints(2, 9))
+    [2, 4, 6, 8, 9]
+    >>> list(_endpoints(2, 2))
+    [2]
+
+    Yields:
+        Values in the range, from `step` and up to `end`.
     """
-    return any(path.is_relative_to(ignore_path) for ignore_path in ignore_paths)
+    yield from range(step, end, step)
+    yield end
 
 
-class FilesSerializer(serialization.Serializer):
-    """Generic file serializer.
+class ShardedFilesSerializer(serialization.Serializer):
+    """Generic file shard serializer.
 
     Traverses the model directory and creates digests for every file found,
-    possibly in parallel.
+    sharding the file in equal shards and computing the digests in parallel.
 
     Subclasses can then create a manifest with these digests, either listing
-    them item by item, or combining everything into a single digest.
+    them item by item, combining them into file digests, or combining all of
+    them into a single digest.
     """
 
     def __init__(
         self,
-        file_hasher_factory: Callable[[pathlib.Path], file.FileHasher],
+        sharded_hasher_factory: Callable[
+            [pathlib.Path, int, int], file.ShardedFileHasher
+        ],
         *,
         max_workers: Optional[int] = None,
         allow_symlinks: bool = False,
@@ -113,17 +95,25 @@ class FilesSerializer(serialization.Serializer):
         """Initializes an instance to serialize a model with this serializer.
 
         Args:
-            file_hasher_factory: A callable to build the hash engine used to
-              hash individual files.
+            sharded_hasher_factory: A callable to build the hash engine used to
+              hash every shard of the files in the model. Because each shard is
+              processed in parallel, every thread needs to call the factory to
+              start hashing. The arguments are the file, and the endpoints of
+              the shard.
             max_workers: Maximum number of workers to use in parallel. Default
               is to defer to the `concurrent.futures` library.
             allow_symlinks: Controls whether symbolic links are included. If a
               symlink is present but the flag is `False` (default) the
               serialization would raise an error.
         """
-        self._hasher_factory = file_hasher_factory
+        self._hasher_factory = sharded_hasher_factory
         self._max_workers = max_workers
         self._allow_symlinks = allow_symlinks
+
+        # Precompute some private values only once by using a mock file hasher.
+        # None of the arguments used to build the hasher are used.
+        hasher = sharded_hasher_factory(pathlib.Path(), 0, 1)
+        self._shard_size = hasher.shard_size
 
     @override
     def serialize(
@@ -147,56 +137,81 @@ class FilesSerializer(serialization.Serializer):
             ValueError: The model contains a symbolic link, but the serializer
               was not initialized with `allow_symlinks=True`.
         """
-        paths = []
+        shards = []
         # TODO: github.com/sigstore/model-transparency/issues/200 - When
         # Python3.12 is the minimum supported version, the glob can be replaced
         # with `pathlib.Path.walk` for a clearer interface, and some speed
         # improvement.
         for path in itertools.chain((model_path,), model_path.glob("**/*")):
-            check_file_or_directory(path, allow_symlinks=self._allow_symlinks)
-            if path.is_file() and not _ignored(path, ignore_paths):
-                paths.append(path)
+            serialize_by_file.check_file_or_directory(
+                path, allow_symlinks=self._allow_symlinks
+            )
+            if path.is_file() and not serialize_by_file._ignored(
+                path, ignore_paths
+            ):
+                shards.extend(self._get_shards(path))
 
         manifest_items = []
         with concurrent.futures.ThreadPoolExecutor(
             max_workers=self._max_workers
         ) as tpe:
             futures = [
-                tpe.submit(self._compute_hash, model_path, path)
-                for path in paths
+                tpe.submit(self._compute_hash, model_path, path, start, end)
+                for path, start, end in shards
             ]
             for future in concurrent.futures.as_completed(futures):
                 manifest_items.append(future.result())
 
         return self._build_manifest(manifest_items)
 
+    def _get_shards(
+        self, path: pathlib.Path
+    ) -> list[tuple[pathlib.Path, int, int]]:
+        """Determines the shards of a given file path."""
+        shards = []
+        path_size = path.stat().st_size
+        if path_size > 0:
+            start = 0
+            for end in _endpoints(self._shard_size, path_size):
+                shards.append((path, start, end))
+                start = end
+        return shards
+
     def _compute_hash(
-        self, model_path: pathlib.Path, path: pathlib.Path
-    ) -> manifest.FileManifestItem:
+        self, model_path: pathlib.Path, path: pathlib.Path, start: int, end: int
+    ) -> manifest.ShardedFileManifestItem:
         """Produces the manifest item of the file given by `path`.
 
         Args:
             model_path: The path to the model.
             path: Path to the file in the model, that is currently transformed
               to a manifest item.
+            start: The start offset of the shard (included).
+            end: The end offset of the shard (not included).
 
         Returns:
             The itemized manifest.
         """
         relative_path = path.relative_to(model_path)
-        digest = self._hasher_factory(path).compute()
-        return manifest.FileManifestItem(path=relative_path, digest=digest)
+        digest = self._hasher_factory(path, start, end).compute()
+        return manifest.ShardedFileManifestItem(
+            path=relative_path, digest=digest, start=start, end=end
+        )
 
     @abc.abstractmethod
     def _build_manifest(
-        self, items: Iterable[manifest.FileManifestItem]
+        self, items: Iterable[manifest.ShardedFileManifestItem]
     ) -> manifest.Manifest:
-        """Builds the manifest representing the serialization of the model."""
+        """Builds an itemized manifest from a given list of items.
+
+        Every subclass needs to implement this method to determine the format of
+        the manifest.
+        """
         pass
 
 
-class ManifestSerializer(FilesSerializer):
-    """Model serializer that produces an itemized manifest, at file level.
+class ManifestSerializer(ShardedFilesSerializer):
+    """Model serializers that produces an itemized manifest, at shard level.
 
     Since the manifest lists each item individually, this will also enable
     support for incremental updates (to be added later).
@@ -235,6 +250,6 @@ class ManifestSerializer(FilesSerializer):
 
     @override
     def _build_manifest(
-        self, items: Iterable[manifest.FileManifestItem]
+        self, items: Iterable[manifest.ShardedFileManifestItem]
     ) -> manifest.Manifest:
         return manifest.Manifest(items)
