+++ src/model_signing/hash.py
@@ -33,8 +33,8 @@ from model_signing import manifest
 from model_signing._hashing import file
 from model_signing._hashing import hashing
 from model_signing._hashing import memory
-from model_signing.serialization import serialize_by_file
-from model_signing.serialization import serialize_by_file_shard
+from model_signing._serialization import serialize_by_file
+from model_signing._serialization import serialize_by_file_shard
 
 
 if sys.version_info >= (3, 11):
