+++ tests/_serialization/testdata/serialize_by_file/TestManifestSerializer/deep_model_folder
