+++ tests/_serialization/testdata/serialize_by_file/TestManifestSerializer/empty_model_file
