+++ tests/_serialization/testdata/serialize_by_file/TestManifestSerializer/model_folder_with_empty_file
