+++ tests/_serialization/testdata/serialize_by_file/TestManifestSerializer/sample_model_file
