+++ tests/_serialization/testdata/serialize_by_file/TestManifestSerializer/symlink_model_folder
