+++ tests/_serialization/testdata/serialize_by_file_shard/TestManifestSerializer/deep_model_folder
