+++ tests/_serialization/testdata/serialize_by_file_shard/TestManifestSerializer/sample_model_folder
