+++ tests/signing/in_toto_signature_test.py
@@ -16,8 +16,8 @@ import pathlib
 
 from model_signing._hashing import file
 from model_signing._hashing import memory
-from model_signing.serialization import serialize_by_file
-from model_signing.serialization import serialize_by_file_shard
+from model_signing._serialization import serialize_by_file
+from model_signing._serialization import serialize_by_file_shard
 from model_signing.signature import fake
 from model_signing.signing import in_toto
 from model_signing.signing import in_toto_signature
