+++ tests/signing/in_toto_test.py
@@ -28,8 +28,8 @@ import pytest
 
 from model_signing._hashing import file
 from model_signing._hashing import memory
-from model_signing.serialization import serialize_by_file
-from model_signing.serialization import serialize_by_file_shard
+from model_signing._serialization import serialize_by_file
+from model_signing._serialization import serialize_by_file_shard
 from model_signing.signing import in_toto
 from tests import test_support
